refactor: 공통을 사용을 위한 디자인 시스템 패키지로 이동함 또한 파일 이름 수정
