+++ presenter/src/main/java/com/mashup/twotoo/presenter/designsystem/component/dialog/selection/SelectionDialogButtonContent.kt
@@ -1,23 +1,23 @@
-package com.mashup.twotoo.presenter.history.model
+package com.mashup.twotoo.presenter.designsystem.component.dialog.selection
 
 import androidx.annotation.StringRes
 import androidx.compose.ui.graphics.Color
 import com.mashup.twotoo.presenter.R
 import com.mashup.twotoo.presenter.designsystem.theme.TwotooPink
 
-data class DropDialogTextUiModel(
+data class SelectionDialogButtonContent(
     @StringRes val titleId: Int,
     val buttonAction: () -> Unit,
     val color: Color,
 ) {
     companion object {
         val default = listOf(
-            DropDialogTextUiModel(
+            SelectionDialogButtonContent(
                 titleId = R.string.challenge_done,
                 {},
                 TwotooPink,
             ),
-            DropDialogTextUiModel(titleId = R.string.cancel, {}, Color.Black),
+            SelectionDialogButtonContent(titleId = R.string.cancel, {}, Color.Black),
         )
     }
 }
