+++ presenter/src/main/java/com/mashup/twotoo/presenter/history/HistoryScreen.kt
@@ -34,11 +34,11 @@ import com.mashup.twotoo.presenter.designsystem.component.bottomsheet.BottomShee
 import com.mashup.twotoo.presenter.designsystem.component.bottomsheet.TwoTooBottomSheet
 import com.mashup.twotoo.presenter.designsystem.component.dialog.DialogContent
 import com.mashup.twotoo.presenter.designsystem.component.dialog.TwoTooDialog
+import com.mashup.twotoo.presenter.designsystem.component.dialog.selection.SelectionDialogButtonContent
 import com.mashup.twotoo.presenter.designsystem.component.loading.FlowerLoadingIndicator
 import com.mashup.twotoo.presenter.designsystem.component.toolbar.TwoTooBackToolbar
 import com.mashup.twotoo.presenter.designsystem.theme.TwoTooTheme
 import com.mashup.twotoo.presenter.designsystem.theme.TwotooPink
-import com.mashup.twotoo.presenter.history.model.DropDialogTextUiModel
 import com.mashup.twotoo.presenter.home.TwoTooGoalAchievementProgressbar
 import com.mashup.twotoo.presenter.home.model.HomeGoalAchievePartnerAndMeUiModel
 import org.orbitmvi.orbit.compose.collectAsState
@@ -174,9 +174,9 @@ fun HistoryScreen(
             }
         }
         if (showSelectListDialog) {
-            ChallengeDropSelectionDialog(
-                dropDialogTextUiModels = listOf(
-                    DropDialogTextUiModel(
+            TwoTooSelectionDialog(
+                selectionDialogButtonContents = listOf(
+                    SelectionDialogButtonContent(
                         titleId = challengeDropSelectionDialogTitleId,
                         buttonAction = {
                             showSelectListDialog = false
@@ -184,7 +184,7 @@ fun HistoryScreen(
                         },
                         color = TwotooPink,
                     ),
-                    DropDialogTextUiModel(
+                    SelectionDialogButtonContent(
                         titleId = R.string.cancel,
                         buttonAction = { showSelectListDialog = false },
                         color = Color.Black,
